fix bug in symlink function

+++ setup.sh
@@ -44,9 +44,8 @@ ask() {
 
 # Symlink helper function
 symlink() {
-  # If this isnt a symlink
-  # if [ ! -L "$2" && -f "$2" ]; then
-  if [ ! -L "$2" ]; then
+  # If this isnt a symlink (and the file exists)
+  if [ ! -L "$2" ] && [ -f "$2" ]; then
     # Backup the file
     mv "$2" "$2.backup" >/dev/null 2>&1
     echo -e "$2 --> $2.backup"
